+++ Blaise.Dqs.Tests.Behaviour/Steps/DeployQuestionnaireSteps.cs
@@ -2,7 +2,7 @@
 using Blaise.Tests.Helpers.Case;
 using Blaise.Tests.Helpers.Configuration;
 using Blaise.Tests.Helpers.Dqs;
-using Blaise.Tests.Helpers.Instrument;
+using Blaise.Tests.Helpers.Questionnaire;
 using NUnit.Framework;
 using System;
 using System.Linq;
@@ -32,7 +32,7 @@ namespace Blaise.Dqs.Tests.Behaviour.Steps
         [Given(@"there is a questionnaire installed in Blaise")]
         public void GivenThereIsAQuestionnaireInstalledInBlaise()
         {
-            InstrumentHelper.GetInstance().InstallInstrument();
+            QuestionnaireHelper.GetInstance().InstallInstrument();
         }
 
         [Then(@"I am presented with a list of the questionnaires already deployed to Blaise")]
@@ -132,12 +132,12 @@ namespace Blaise.Dqs.Tests.Behaviour.Steps
         [Given(@"I have been presented with questionnaire already exists screen")]
         public void GivenIHaveBeenPresentedWithQuestionnaireAlreadyExistsScreen()
         {
-            InstrumentHelper.GetInstance().InstallInstrument();
+            QuestionnaireHelper.GetInstance().InstallInstrument();
             DqsHelper.GetInstance().LoadUploadPage();
             DqsHelper.GetInstance().SelectQuestionnairePackage();
             DqsHelper.GetInstance().ConfirmQuestionnaireUpload();
 
-            _scenarioContext.Set(InstrumentHelper.GetInstance().GetInstallDate(), InstallDate);
+            _scenarioContext.Set(QuestionnaireHelper.GetInstance().GetInstallDate(), InstallDate);
         }
 
         [When(@"I select to cancel")]
@@ -163,7 +163,7 @@ namespace Blaise.Dqs.Tests.Behaviour.Steps
         public void ThenTheQuestionnaireHasNotBeenOverwritten()
         {
             var expectedInstallDate = _scenarioContext.Get<DateTime>(InstallDate);
-            var actualInstallDate = InstrumentHelper.GetInstance().GetInstallDate();
+            var actualInstallDate = QuestionnaireHelper.GetInstance().GetInstallDate();
 
             Assert.AreEqual(expectedInstallDate, actualInstallDate);
         }
@@ -181,7 +181,7 @@ namespace Blaise.Dqs.Tests.Behaviour.Steps
         {
             DqsHelper.GetInstance().WaitForUploadToComplete();
             var existingInstallDate = _scenarioContext.Get<DateTime>(InstallDate);
-            var newInstallDate = InstrumentHelper.GetInstance().GetInstallDate();
+            var newInstallDate = QuestionnaireHelper.GetInstance().GetInstallDate();
 
             Assert.Greater(newInstallDate, existingInstallDate);
         }
@@ -189,24 +189,24 @@ namespace Blaise.Dqs.Tests.Behaviour.Steps
         [Then(@"the questionnaire is active in blaise")]
         public void ThenTheQuestionnaireIsActiveInBlaise()
         {
-            var instrumentInstalled = InstrumentHelper.GetInstance().SurveyHasInstalled(BlaiseConfigurationHelper.InstrumentName, 60);
+            var instrumentInstalled = QuestionnaireHelper.GetInstance().SurveyHasInstalled(BlaiseConfigurationHelper.InstrumentName, 60);
             Assert.IsTrue(instrumentInstalled);
         }
 
         [Given(@"the package I have selected already exists in Blaise")]
         public void GivenThePackageIHaveSelectedAlreadyExistsInBlaise()
         {
-            InstrumentHelper.GetInstance().InstallInstrument();
+            QuestionnaireHelper.GetInstance().InstallInstrument();
         }
 
         [AfterScenario("questionnaire")]
         public void CleanUpScenario()
         {
             DqsHelper.GetInstance().LogOutOfToDqs();
-            if (InstrumentHelper.GetInstance().SurveyExists(BlaiseConfigurationHelper.InstrumentName))
+            if (QuestionnaireHelper.GetInstance().SurveyExists(BlaiseConfigurationHelper.InstrumentName))
             {
                 CaseHelper.GetInstance().DeleteCases();
-                InstrumentHelper.GetInstance().UninstallSurvey();
+                QuestionnaireHelper.GetInstance().UninstallSurvey();
             }
             BrowserHelper.ClosePreviousTab();
         }
