+++ Blaise.Tests.Behaviour/Features/CreateCases.feature
@@ -8,7 +8,7 @@ Feature: Create Cases
 @regression
 Scenario: Create sample cases to capture respondent data
 	Given I have an questionnaire installed on a Blaise environment
-	When I create sample cases for the instrument
+	When I create sample cases for the questionnaire
 		| primarykey | outcomecode | telephoneno  |
 		| 900001     | 110         | 07000 000 00 |
 		| 900002     | 110         | 07000 000 00 |
