+++ Blaise.Tests.Behaviour/Steps/InstallQuestionnaireSteps.cs
@@ -1,7 +1,7 @@
 ﻿using Blaise.Nuget.Api.Contracts.Enums;
 using Blaise.Tests.Helpers.Browser;
 using Blaise.Tests.Helpers.Configuration;
-using Blaise.Tests.Helpers.Instrument;
+using Blaise.Tests.Helpers.Questionnaire;
 using NUnit.Framework;
 using TechTalk.SpecFlow;
 
@@ -28,19 +28,19 @@ namespace Blaise.Tests.Behaviour.Steps
         [When(@"I install the questionnaire into a Blaise environment")]
         public void WhenIInstallTheInstrumentIntoABlaiseEnvironment()
         {
-            InstrumentHelper.GetInstance().InstallInstrument();
+            QuestionnaireHelper.GetInstance().InstallInstrument();
         }
 
         [When(@"I install the questionnaire into a Blaise environment specifying a Cati configuration")]
         public void WhenIInstallTheInstrumentIntoABlaiseEnvironmentSpecifyingACatiConfiguration()
         {
-            InstrumentHelper.GetInstance().InstallInstrument();
+            QuestionnaireHelper.GetInstance().InstallInstrument();
         }
 
         [Then(@"the questionnaire is available to use in the Blaise environment")]
         public void ThenTheInstrumentIsAvailableToUseInTheBlaiseEnvironment()
         {
-            var instrumentHasInstalled = InstrumentHelper.GetInstance().SurveyHasInstalled(BlaiseConfigurationHelper.InstrumentName, 60);
+            var instrumentHasInstalled = QuestionnaireHelper.GetInstance().SurveyHasInstalled(BlaiseConfigurationHelper.InstrumentName, 60);
 
             Assert.IsTrue(instrumentHasInstalled, "The instrument has not been installed, or is not active");
         }
@@ -48,14 +48,14 @@ namespace Blaise.Tests.Behaviour.Steps
         [Then(@"the questionnaire is configured to capture respondents data via Cati")]
         public void ThenTheInstrumentIsConfiguredToCaptureRespondentsDataViaCati()
         {
-            var surveyConfiguration = InstrumentHelper.GetInstance().GetSurveyInterviewType();
+            var surveyConfiguration = QuestionnaireHelper.GetInstance().GetSurveyInterviewType();
             Assert.AreEqual(QuestionnaireInterviewType.Cati, surveyConfiguration);
         }
 
         [AfterScenario("instrument")]
         public void CleanUpScenario()
         {
-            InstrumentHelper.GetInstance().UninstallSurvey();
+            QuestionnaireHelper.GetInstance().UninstallSurvey();
         }
 
         [AfterTestRun]
